+++ lib/karafka/routing/features/deserializers.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Karafka
+  module Routing
+    module Features
+      class Deserializers < Base
+      end
+    end
+  end
+end