+++ spec/lib/karafka/messages/builders/batch_metadata_spec.rb
@@ -21,7 +21,7 @@ RSpec.describe_current do
     it { expect(result.first_offset).to eq kafka_batch.first.offset }
     it { expect(result.last_offset).to eq kafka_batch.last.offset }
     it { expect(result.topic).to eq routing_topic.name }
-    it { expect(result.deserializer).to eq routing_topic.deserializer }
+    it { expect(result.deserializers).to eq routing_topic.deserializers }
     it { expect(result.scheduled_at).to eq(scheduled_at) }
 
     it 'expect to have processed_at set to nil, since not yet picked up' do
@@ -44,7 +44,7 @@ RSpec.describe_current do
       it { expect(result.first_offset).to eq(-1001) }
       it { expect(result.last_offset).to eq(-1001) }
       it { expect(result.topic).to eq routing_topic.name }
-      it { expect(result.deserializer).to eq routing_topic.deserializer }
+      it { expect(result.deserializers).to eq routing_topic.deserializers }
       it { expect(result.scheduled_at).to eq(scheduled_at) }
     end