+++ spec/lib/karafka/messages/message_spec.rb
@@ -10,7 +10,9 @@ RSpec.describe_current do
     let(:deserializer) { ->(_) { 1 } }
     let(:metadata) do
       ::Karafka::Messages::Metadata.new.tap do |metadata|
-        metadata['deserializer'] = deserializer
+        metadata['deserializers'] = Karafka::Routing::Features::Deserializers::Config.new(
+          payload: deserializer
+        )
       end
     end