+++ spec/support/factories/messages/metadata.rb
@@ -7,11 +7,18 @@ FactoryBot.define do
     sequence(:topic) { |nr| "topic-from-meta#{nr}" }
     sequence(:offset) { |nr| nr }
     partition { 0 }
-    deserializer { ->(message) { JSON.parse(message.raw_payload) } }
     timestamp { Time.now.utc }
+    deserializers do
+      Karafka::Routing::Features::Deserializers::Config.new(
+        active: true,
+        payload: ->(message) { JSON.parse(message.raw_payload) },
+        key: Karafka::Serialization::Passthrough::Deserializer.new,
+        headers: Karafka::Serialization::Passthrough::Deserializer.new
+      )
+    end
 
     initialize_with do
-      new(partition: partition, topic: topic, offset: 0, deserializer: deserializer)
+      new(partition: partition, topic: topic, offset: 0, deserializers: deserializers)
     end
   end
 end